+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # zhinst-toolkit Changelog
 
 ## Version 0.4.4
-* Added full support for the LabOne Scope module (no need to fallback to zhinst.core)
+* Added full support for the following LabOne modules (no need to fallback to zhinst.core):
+  * Scope Module
+  * Sweeper Module
 * Renamed `zhinst.toolkit.nodetree.nodes.WildcardResult` to `zhinst.toolkit.nodetree.helpers.NodeDict`
 
 ## Version 0.4.3